+++ src/client/GraphQLClient/types/client.ts
@@ -1,4 +1,4 @@
-import { TokenPermissions } from '../../../types';
+import { RepoKeyAndName, TokenPermissions } from '../../../types';
 
 /**
  * From the name given to the data for the JupiterOne use case to the specific Github
@@ -103,6 +103,20 @@ export interface OrgRepoQueryResponse extends Node {
   isArchived: boolean | undefined;
   createdAt: string;
   updatedAt: string;
+  node_id?: string;
+}
+
+export interface OrgSecretRepoQueryResponse {
+  id: number;
+  node_id: string;
+  name: string;
+  full_name: string;
+  owner?: object;
+  private: boolean;
+  html_url: string;
+  description: string;
+  fork: boolean;
+  url: string;
 }
 
 export interface OrgTeamRepoQueryResponse extends OrgRepoQueryResponse {
@@ -111,7 +125,7 @@ export interface OrgTeamRepoQueryResponse extends OrgRepoQueryResponse {
 }
 
 export interface OrgAppQueryResponse {
-  //a REST response, not GraphQL, but everything else is in this file
+  //a REST response, not GraphQL
   id: string; //the installation id
   respository_selection: string;
   html_url: string;
@@ -129,6 +143,20 @@ export interface OrgAppQueryResponse {
   single_file_paths: string[];
   suspended_by: string;
   suspended_at: string;
+}
+
+export interface OrgSecretQueryResponse {
+  //a REST response, not GraphQL
+  name: string;
+  created_at: string;
+  updated_at: string;
+  visibility?: string; // 'private' | 'all' | 'selected'. This means how many repos can use this secret
+  selected_repositories_url?: string; //a webpage url, not a REST API url
+  //the following fields are set by the integration code, not received from the REST API
+  orgLogin?: string; //for use in constructing weblinks
+  secretOwnerType?: string; // 'org' | 'repo' | 'env'
+  secretOwnerName?: string;
+  repos?: RepoKeyAndName[]; //to help build relationships with minimal memory footprint, using RepoKeyAndName instead of RepoEntity
 }
 
 interface GithubResources {