+++ src/steps/index.ts
@@ -5,6 +5,8 @@ import { memberSteps } from './members';
 import { prSteps } from './pullrequests';
 import { repoSteps } from './repos';
 import { teamSteps } from './teams';
+import { orgSecretSteps } from './orgsecrets';
+import { repoSecretSteps } from './reposecrets';
 
 const integrationSteps = [
   ...accountSteps,
@@ -14,6 +16,8 @@ const integrationSteps = [
   ...collaboratorSteps,
   ...prSteps,
   ...appSteps,
+  ...orgSecretSteps,
+  ...repoSecretSteps,
 ];
 
 export { integrationSteps };